src/requirements.txt
Flask
pymysql
python-dotenv==1.0.0
